create useUsers hook and H1 and H2 generic components
